Create scaleRegressAndProjectPCA.R
This is probably a pretty confusing function, but it takes a seurat object and scales it in the specified way (specify vars.to.regress) using a reference seurat object. This is to ensure that when we layer on top, the new seurat object undergoes the same exact transformation that the old one underwent.